+++ source/renderer/app/containers/LoadingPage.js
@@ -9,6 +9,9 @@ import cardanoLogo from '../assets/images/cardano-logo.inline.svg';
 import type { InjectedProps } from '../types/injectedPropsType';
 import { generateFileNameWithTimestamp } from '../../../common/utils/files';
 import { getSupportUrl } from '../utils/network';
+import NotificationMessage from '../components/widgets/NotificationMessage';
+import successIcon from '../assets/images/success-small.inline.svg';
+import { DOWNLOAD_LOGS_SUCCESS_DURATION } from '../config/timingConfig';
 
 export const messages = defineMessages({
   loadingWalletData: {
@@ -21,6 +24,11 @@ export const messages = defineMessages({
     defaultMessage: '!!!https://iohk.zendesk.com/hc/en-us/requests/new/',
     description: 'Link to Open Support page'
   },
+  downloadLogsSuccess: {
+    id: 'loading.screen.reportIssue.downloadLogsSuccessMessage',
+    defaultMessage: '!!!Logs successfully downloaded',
+    description: 'Success message for download logs.',
+  },
 });
 
 @inject('stores', 'actions') @observer
@@ -29,6 +37,18 @@ export default class LoadingPage extends Component<InjectedProps> {
   static contextTypes = {
     intl: intlShape.isRequired,
   };
+
+  constructor(props: any, context: any) {
+    super(props);
+    this.context = context;
+    this.registerOnDownloadLogsNotification();
+  }
+
+  componentWillUnmount() {
+    const { profile } = this.props.actions;
+    profile.downloadLogs.remove(this.openNotification);
+    this.closeNotification();
+  }
 
   render() {
     const { stores } = this.props;
@@ -39,6 +59,7 @@ export default class LoadingPage extends Component<InjectedProps> {
       isNotEnoughDiskSpace, diskSpaceRequired, diskSpaceMissing, diskSpaceRecommended,
     } = stores.networkStatus;
     const { hasLoadedCurrentLocale, hasLoadedCurrentTheme, currentLocale } = stores.profile;
+    const { id, message } = this.notification;
     return (
       <CenteredLayout>
         <Loading
@@ -66,9 +87,32 @@ export default class LoadingPage extends Component<InjectedProps> {
           onContinueWithoutClockSyncCheck={ignoreSystemTimeChecks}
           onDownloadLogs={this.handleDownloadLogs}
         />
+        <NotificationMessage
+          icon={successIcon}
+          show={stores.uiNotifications.isOpen(id)}
+          onClose={this.closeNotification}
+          clickToClose
+          hasCloseButton
+        >
+          {message}
+        </NotificationMessage>
       </CenteredLayout>
     );
   }
+
+  get notification() {
+    const { intl } = this.context;
+    return {
+      id: 'loading-page-download-logs-success',
+      duration: DOWNLOAD_LOGS_SUCCESS_DURATION,
+      message: intl.formatMessage(messages.downloadLogsSuccess),
+    };
+  }
+
+  closeNotification = () => {
+    const { id } = this.notification;
+    this.props.actions.notifications.closeActiveNotification.trigger({ id });
+  };
 
   handleReportIssueClick = async (event: SyntheticEvent<HTMLButtonElement>) => {
     event.persist();
@@ -79,13 +123,25 @@ export default class LoadingPage extends Component<InjectedProps> {
     this.props.stores.app.openExternalLink(supportUrl);
   };
 
+  openNotification = () => {
+    const { notifications } = this.props.actions;
+    const { id, duration } = this.notification;
+    notifications.open.trigger({ id, duration });
+  };
+
+  registerOnDownloadLogsNotification = () => {
+    const { profile } = this.props.actions;
+    profile.downloadLogs.listen(this.openNotification);
+  };
+
   handleDownloadLogs = () => {
+    const { profile } = this.props.actions;
     const fileName = generateFileNameWithTimestamp();
     const destination = global.dialog.showSaveDialog({
       defaultPath: fileName,
     });
     if (destination) {
-      this.props.actions.profile.downloadLogs.trigger({ fileName, destination, fresh: true });
+      profile.downloadLogs.trigger({ fileName, destination, fresh: true });
     }
   };
 }